requirements.txt
aniso8601==6.0.0
attrs==19.3.0
certifi==2019.9.11
chardet==3.0.4
Django==2.2.4
django-filter==2.1.0
django-graphiql==0.4.4
djangorestframework==3.10.3
graphene==2.1.6
graphene-django==2.3.2
graphene-django-extras==0.4.8
graphql-core==2.2
graphql-relay==0.4.5
idna==2.8
importlib-metadata==0.23
more-itertools==7.2.0
packaging==19.2
pluggy==0.13.0
promise==2.2.1
py==1.8.0
pyparsing==2.4.5
pytest==5.3.0
python-dateutil==2.8.1
pytz==2019.1
requests==2.22.0
Rx==1.6.1
singledispatch==3.4.0.3
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.7
wcwidth==0.1.7
zipp==0.6.0
